compute-client: better consolidate introspection updates
This commit reorders controller maintenance steps to achieve a better
consolidation of introspection updates persisted to storage. Some of the
maintenance steps can result in dropping of controller state, which in
turn can result in retraction of introspection data, so if we run those
steps before the introspection reporting, chances are more introspection
updates consolidate away and persist writes become smaller.