compute-client: fix replica write frontier initialization
When installing write frontiers for a new replica, we should not attempt
to install empty read holds for dataflow dependencies with the storage
controller. If a collection has advanced to the empty frontier, this
means we don't have read holds on its inputs anymore, so the storage
controller has been free to drop these input collections. If we try to
update the storage read capabilities for an object the storage
controller doesn't know about, it panics, so we must avoid that.