catalog: Fix persist savepoint updates
When processing updates in the persist catalog, retractions need to be
processed before insertions for a given timestamp, so we retract the
old value and not the newly inserted value. The savepoint persist
catalog was not sorting updates properly for a given timestamp and not
processing updates correctly. This commit fixes the issue by sorting
updates within a timestamp.

Fixes #25363