dyncfg: extract list of all dyncfgs in the system into a new crate
This deduplicates the two existing copies into one.

On on hand, it seems silly to have a whole crate just for this. On the
other hand, there are some tricky dependency issues involved.

For dyncfgs to work, this list needs to be the same in both `clusterd`
and `environmentd`. (Technically, I think it would be okay for the
`clusterd` list to be a subset of the `environmentd` list, but that's
pretty subtle to reason about to let's avoid it if possible.) This means
we need to find a home for it that can depend on all crates that use
dyncfgs (initially persist-client and persist-txn, but hopefully more
soon) and that can be depended on by both `clusterd` and `environmentd`
main files. The initial location (in mz-sql) already fails the latter
requirement, which is why it had to be duplicated.

It's possible that we could find a crate that happens to meet these
requirements now, but as we add more uses of dyncfgs, it seems like that
would get increasingly harder. So, rip off the bandaid now and make a
crate specifically for this, I guess.

They hypothesized link-time magic registry would sidestep all this, but
it has some other footguns we're working through.

While I'm in here, rename the methods to `all_dyncfgs` to be more
consistent with the "dyncfg" jargon.