adapter: Remove redundant message count metric (#25389)
This commit removes the mz_coord_messages_processed metric that counts
the total number of messages processed by the Coordinator. This metric
was added in 57a56d6f46681ec25cf5592930396a2f4c70c5cf. It is redundant
with the slow_message_handling metric. As of
02bb658a445ffb87d3c8d28872f2aee08c5f2144, slow_message_handling counts
all messages not just the slow ones.

Additionally, the commit removes the watchdog_messages_processed metric
that counts the total number of watchdog messages processed by the
Coordinator. It's replaced by a new label in the slow_message_handling
metric called "watchdog".