persist-txn: fix "register ts ... not less_equal forget ts" bug
This debug assertion was getting tickled by the following sequence of
events:

- At time `A`, a data shard is registered
- At time `B`, it is written to via txns
- At time `C`, it is forgotten
- We then open the txns cache with an as_of time `D`
- We read the register and write into the cache. The pushing the write
  into the cache triggers a call to `compact_data_times`, which advances
  the register ts from `A` to `D`
- We then read the forget into the cache, the advanced register ts `D`
  is not less_equal the forget ts `C` and boom

I can't think of any reason for us to advance the register ts. It's
likely a holdover from before we kept around the original timestamps of
things. So, fix the bug by removing the advancement from `A` -> `C`,
leaving the register ts in the cache at `A`.

Closes #25384