Expression analysis framework (#25292)
This PR introduces an `Analysis` trait that is meant to be a
simplification of the `Attribute` trait/framework.

It is simpler in many ways, and less expressive, but seemingly the
restrictions are not yet limiting and are potentially helpful in
broadening the uptake of the framework.

The main idea is that an `Analysis` implementor specifies an output type
it will produce for each expression, and the logic that produces this
output given an expression, analysis results for child expressions, and
analysis results for other depended-upon analyses. All analysis state is
indexed by "offset": the post-order visitation numbering of the
expressions.

The framework runs the analyses in dependency order, establishes links
between `Let`-bound identifiers and offsets, but otherwise has very few
opinions. By contrast, the Attribute framework allowed its hosted
attributes to propose various bits of logic that would run in addition
as it traverses an expression, used to set up `Let`-binding information
but also potentially much more.

There is a demonstration of how to use the framework in
https://github.com/MaterializeInc/materialize/pull/24155/, but we are
looking at this PR in isolation to break up the moving parts of that PR
into manageable pieces.
1. The main annoyance in using the output is that one must keep notes as
one descends into an expression about the current expression offsets, as
very little help is currently offered here (though, a `Slice` variant of
the results that scopes itself down to expressions, and their child
expressions, would probably be helpful).
2. There is a fair amount of cloning of results, even though many
expressions only wish to say "whatever my only child says". One could
imagine an analysis result type of `Result<AnalysisType, usize>` where
the `Err(offset)` variant might indicate an offset to consult for the
state (and expressions that don't modify the analysis result could have
`Err(offset)` for some appropriate `offset` (their child, or whereever
their child references).

The main reviewing question is whether this simplification is net
valuable, in that it makes things easier to use and adopt, without
ruling out too much expressivity or potentially performance.
Additionally, I have not yet begun to integrate it further than the
previous PR, but I'm happy to follow up with commits that do this if we
identify where we could most fruitfully slice out the existing uses. I
put together prototype versions for all existing attributes other than
`Cardinality` and `ColumnNames`, as they each had a fair bit of inherent
complexity.

### Motivation

<!--
Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug.

    [Ensure issue is linked somewhere.]

  * This PR adds a known-desirable feature.

    [Ensure issue is linked somewhere.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

[Write a brief specification for the feature, including justification
for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

[Describe what was wrong with the existing code, if it is not obvious.]
-->

### Tips for reviewer

<!--
Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
* [Some function/module/file] is pure code movement and only needs a
skim.

Delete this section if no tips.
-->

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->