sql: Fix timestamp precision in HTTP API
Previously, the HTTP and WS API would return TIMESTAMPs and
TIMESTAMPTZs with only millisecond precision, even though internally
and via the pgwire API we represent these types with microsecond
precision. This commit updates the HTTP and WS API to return
microsecond precision.

To avoid issues of compatibility with current users, the microsecond
portion of the timestamp is stored as a decimal. For example,
previously `TIMESTAMP '2023-12-19T06:50:37.123456 '` would return
`"1703011837123"`, now it returns `"1703011837123.456"`.

An alternative implementation that used arbitrary precision decimals
(the `Numeric` type in mz_repr) was explored, but the performance was
2-4x slower than the current implementation.

Fixes #24026