persist: remove allocations in Codec decode codepath
In the (presumably) common case of Rows made entirely of stack-allocated
`Datum` (`Bool`, `Int64`, etc) this reduces the allocations per Row from
2 to 0. `Bytes` could be the same with careful use of prost's
`bytes::Bytes` feature. `String` would need something like Parker's
https://github.com/MaterializeInc/prost/pull/6. For more complex types
(`List`, `AclItem`, etc) this PR is probably a rounding error.

This is accomplished by adding a new `decode_into` variant of the
"legacy" `Codec::decode` method which enables amortizing allocations
across calls in two ways. First, instead of returning `Self`, it takes
`&mut Self` as a parameter, allowing for reuse of the internal
`Row/SourceData` allocations. Second, it adds a `type Storage` to
`Codec` and also passes it in as `&mut Option<Self::Storage>`, allowing
for reuse for ProtoRow/ProtoSourceData allocations. `decode_into` gets a
default implementation for `Codec`s that can't take advantage of this.

This has no effect on the "structured" `Codec` path that we'll hopefully
soon be moving to. I'm generally hesitant to spend too much time
optimizing the "legacy" path, but this is simple enough and a big enough
win that it felt worth it to me.

We then take advantage of this in the `Codec` impls for `Row` (entirely
unused and could be remove) and `SourceData` (the main one used in
production).

To reuse the `Vec<ProtoDatum>` in the `ProtoRow` in the `Ok` variant of
`ProtoSourceData`, we take advantage of prost's `Message::merge`, which
decodes an encoded proto into an existing message using proto's [merge]
semantics. We do this because `Message::merge` seems to be careful and
smart about reusing existing allocations when it can. To make this work
(we don't actually want "merge" semantics), we first clear the message,
so we're merging into an empty one. Unfortunately, the generated code
for `ProtoSourceData::clear` is not able to keep the allocations around,
so we have to manually "clear" it to the equivalent of an `Ok` with an
empty `Row`.

[merge]: https://protobuf.dev/programming-guides/encoding/#last-one-wins

To reuse the internal allocs of the `Row` inside the `SourceData`, we
optimize for the (definitely) common case of `Ok`. This is done manually
with some match statements and by exposing a new `Row::decode_from`
method.

The benchmarks should be no-ops but are not. I have no explanation for
this, but it seems to reproduce.

SourceData

    roundtrip_int64_decode_legacy
                            time:   [1.7787 ms 1.7798 ms 1.7810 ms]
                            change: [-2.2133% -1.8816% -1.5910%] (p = 0.00 < 0.05)
                            Performance has improved.

    roundtrip_bytes_decode_legacy
                            time:   [2.7741 ms 2.7763 ms 2.7800 ms]
                            change: [-0.3923% -0.2916% -0.1545%] (p = 0.00 < 0.05)
                            Change within noise threshold.

    roundtrip_string_decode_legacy
                            time:   [2.4841 ms 2.4855 ms 2.4872 ms]
                            change: [+2.0101% +2.1176% +2.2234%] (p = 0.00 < 0.05)
                            Performance has regressed.

Row

    roundtrip_decode_legacy time:   [1.8818 ms 1.8834 ms 1.8851 ms]
                            change: [-6.4160% -5.9947% -5.6271%] (p = 0.00 < 0.05)
                            Performance has improved.