platform-checks: Fix preflight and rollback scenario
Seen in https://buildkite.com/materialize/nightlies/builds/6561#018dccef-ef30-4f63-824c-ff8d982392d8

Broken by https://github.com/MaterializeInc/materialize/pull/25276/files

This scenario is slightly annoying because it is the only one that ends
up running the verify step in the old version. I'm not sure if this will
be too much effort in the future to maintain.