compute: attach hydration logging operators to arrangements first
If the `CollectionBundle` of a rendered `Plan` node contains an
arrangement, attach the hydration logging operator to that arrangement's
stream rather than the stream of the unarranged collection. The
rationale is that attaching to the unarranged collection will produce
additional forks in the dataflow graph, and therefore unnecessary
cloning of data, which attaching to an arrangement avoids.