fivetran-destination: Specify workspace default-members to make building easier (#25437)
Many folks will run `cargo check` from the root of the workspace to make
sure everything builds successfully. Often they hit an issue where the
`fivetran-destination` fails to build because it depends on the
`/misc/fivetran-sdk` submodule, and then they have to figure out how to
update the submodule which isn't the best experience. The
`fivetran-destination` is touched by very few people (only me at the
moment) so forcing everyone to build it adds unnecessary friction.

The crates specified in `default-members` are the ones that will get
built when running `cargo` commands from the root of our workspace when
a target is not specified. Currently the list of `default-members` are
all binaries that are referenced from `mzbuild.yml` files which should
encompass everything.

Note: This also changes the behavior of `cargo test` when run from the
root of the workspace, it will now only test the crates in
`default-members`. I don't think this is too big of an issue because the
only workflow that requires this is CI. I updated CI to specify the
`--workspace` flag where needed to test the entire workspace.

### Motivation

Improve the developer experience.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a