[explain] Syntax / etc. for EXPLAIN [FILTER] PUSHDOWN (#25053)
This allows parsing statements of the form `EXPLAIN FILTER PUSHDOWN FOR
<explainee>`, where [explainee is as defined in our existing "explain
plan" syntax](https://materialize.com/docs/sql/explain-plan/).

### Motivation

Epic: #24498 

I hear that it's good to try and PR the parser changes separately from
the implementation, so here we go!

### Tips for reviewer

We haven't settled on the exact syntax or output format of this change
yet, so it's hidden by a flag from users.

This is written to accept exactly the set of explainees (SELECTs, MVs)
that regular EXPLAIN supports. It's not entirely clear that we'll
actually end up supporting all of these, but we _might_, and anyways
there's no reason to filter them out in the parser.

Right now this doesn't support any additional formatting or WITH-type
options. Right now we haven't felt the need for any, but if we decide to
add them we should be able to do so in a backwards-compatible way.