adapter: refactor txn dry run
Transaction dry runs were the last use of catalog_transact_with;
to support these, we could instead move the feature into an Op,
allowing us to remove the closure which takes ownership of the
Coordinator.

The last bit is desirable because we might want to provide mutable
access to the storage controller during catalog transactions, which
that reference precludes.

I think this design is also defensible on its own merits as it
broadly simplifies catalog transactions.