compute: move join configuration to dyncfg
This commit ports the two join config options `enable_mz_join_core` and
`linear_join_yielding` to dyncfg.

In contrast to the persist dyncfgs we don't define the compute ones
close to their usage sites because these are in `mz_compute` and we
don't want to require everyone that uses dyncfgs (like adapter) to
depend on that crate. Instead the compute configs are defined in
`mz_compute_types`.