expr,sql: fix `to_jsonb` implementation for arrays
Materialize was not correctly handling arrays in `to_jsonb`. Instead of
converting the arrays to JSON arrays, Materialize was converting them to
a JSON string of the array's textual representation. This behavior was
at variance with PostgreSQL.

The implementation is somewhat irritating, as the size of the array is
not known at plan time. So the planner has to construct an expression to
cast the array's element type and plumb that through to the dataflow
layer, so that each value in the array can be cast to JSON at runtime.

This commit also adjust the implementation for lists in the same way,
for symmetry, even though it's not required for PostgreSQL
compatibility, as lists are a custom Materialize type.

Fix #12181.