adapter: rename "unmanaged" replica to "unorchestrated" replica
Unmanaged replicas are an internal-only feature used for testing
Materialize with finer control than going through the orchestrator would
allow. But the name of the feature clashes badly with the name of a
user-facing feature added later: unmanaged and managed clusters.

So, rename the unmanaged replica feature in the codebase from to
"unorchestrated replicas". It's just a better name overall, as what you
avoid with unorchestrated replicas is Materialize using its
"Orchestrator".