compute: use a growable stack for rendering
This commit makes compute rendering use stacker to grow the stack if
necessary. This is to avoid stack overflows when rendering deeply nested
plans. We don't set a recursion limit because there is currently no way
to gracefully fail rendering of a dataflow.

In addition, this commit bumps the size of stacker's red zone to twice
the previous value. Without this, the query from #25545 would still
cause a stack overflow, apparently because some of the stack frames in
rendering are larger than the previous red zone.