[dbt] dbt deploy (#25174)
### Motivation

Fixes #24186.

We want to integrate blue green deployment work into a simple to use dbt
macro. We introduce a new `deployment` configuration in dbt_project.yml
that are scoped to the current target.

Users specify any number of clusters and schemas to be part of a
deployment.

```yaml
vars:
  deployment:
    default:
      clusters: [usecase1, usecase2, usecase3]
      schemas: [usecase1_subpart1, usecase1_subpart2, usecase2, usecase3]
```

There are then different commands to execute: 

### `dbt run-operation deploy_init`

Automatically creates the deployment schemas and clusters. The naming
convention appends `_dbt_deploy` to the end of each production schema
and cluster that is being clone. Clusters are created with the same
configuration as their corresponding prod target. The command will error
if the `prod` cluster or schema does not exist. It is valid for the
`_dbt_deploy` schema or cluster to be precreated. For example, a user
might want to deploy on a different sized cluster. If the target cluster
or schema does exist, the command will output a warning if it contains
any objects.

### `dbt run --vars 'deploy: True'`

Executing dbt run or build with the deploy flag will intercept the
clusters and schemas for all models and rewrite them with the
_dbt_deploy suffix.

### `dbt run-operation deploy_wait`

This command waits until all clusters are hydrated and have lag less
than 1 second.

### `dbt run-operation deploy_promote(wait=False)`

Executes the alter swap commands in a transaction. It will fail if any
of the clusters or schemas do not exist. By default it does not check
that clusters are hydarated. Users are encouraged to call `deploy_wait`
manually and then run further tests and validation.

### `dbt run-operation deploy_cleanup`

Drops all the `_dbt_deploy` schemas and clusters for easy cleanup. 


### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->

---------

Co-authored-by: Marta Paes <marta.paes.moreira@gmail.com>