[persist] Track the actual upper in validations (#25090)
### Motivation

https://github.com/MaterializeInc/materialize/issues/25015

I'm not confident that this is actually the issue from that ticket, but
I have low enough confidence in the old assertions that I don't feel
like they're especially indicative of a bug, so I think it's worth
updating them regardless.

- The empty antichain is sorted lower than any other antichain, not
higher as you'd expect. (The lower is never the empty antichain, but
this can happen for the upper.)
- The second block of tests would look at the upper of the largest
batch, which is not guaranteed to be the batch with the largest upper.
(Particularly given the above!)

Additionally, the result of the `PartialOrder` call in these tests was
unused. It was unclear to me what the intent of those calls was; I've
removed them here.