catalog: Remove shadow catalog
The shadow catalog was used to test the persist catalog and compare the
results to the stash catalog. Now that the persist catalog has been
rolled out to all environments we no longer need to constantly test the
persist catalog. Removing the shadow catalog will also give us some
speed benefits in CI.

Part of #24218