transform: run local normalization when `JoinFusion` succeeds
Apply `PredicatePushdown` and `CanonicalizeMfp` on the transformed
relation when `fusion::Join` appies a rewrite on the most general
pattern.

This is a necessary (but hopefully temporary fix) to fix regressions
observed in the "Product limits" tests when the PR was originally
tested (https://buildkite.com/materialize/nightlies/builds/6670).

See the Rust comment in the commit diff for more details.