compute: remember topological order in `FlatPlan`
This commit adds an additional field, `topological_order`, to `FlatPlan`
that stores a valid topological order of nodes. It's easy to produce
such an order during `Plan` flattening, so we store it in the plan for
use during rendering. This lets us stop relying on the (maybe wrong,
certainly brittle) assumption that node IDs are always in topological
order.