README: update wording to feature cloud more heavily
We've had several reports of users getting confused by our README and
come away thinking that Materialize is primarily a single node system
with no fault tolerance or horizontal scalability.

This commit is a small reconfiguration of our README to minimize that
confusion.

It also bundles some other small improvements:

  * More links to contact us and join our Slack
  * A fix to the contact us link in the LICENSE file
  * Clarification that `HAVING` is not an aggregate function
  * Use of the correct "good for external contributors" label
  * Less dire warnings about using ORMs, since we've gotten increasingly
    good at supporting ORMs over the years