timely-util: new reclock implementation
Implement a new reclock operator which has asymptotically better
performance than the existing one by taking advantage of a new
technique and a new assumption.

The technique exploited by this implementation is maintaining the pile
of data waiting to be reclocked in runs of updates that are mutually
comparable in their partial order. Maintaining those runs (called
chains[1]) allows for very efficient extraction of new batches of data
as we can leverage binary search within each chain. Furthermore, for all
the concrete timestamp types used for reclocking the linear extension
order implemented by `Ord` leads to an optimal number of chains for a
given subset.

The new assumption exploited by this implementation is assuming that the
timestamp we're reclocking onto is totally ordered. This is by no means
a requirement but mostly a result of the time pressure to release this
to fix #22128. The implementation of reclocking onto a `Lattice` would
result in much more code that would have to review in short notice so it
is left as future work.

[1] https://en.wikipedia.org/wiki/Total_order#Chains

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>