catalog: Move TableTransaction out of stash
This commit moves the TableTransaction struct out of the mz_stash
crate and into the mz_catalog crate. The struct is only ever used by
mz_catalog, and it makes more sense for it to live in that crate.
Especially since we may soon remove the mz_stash crate.

The vast majority of the commit is code movement and the following
concrete changes:

   - Remove `pub` visibility modifier from all `TableTransaction` methods.
   - Change method signatures to return a new variant of
   `DurableCatalogError` instead of `StashError`.
   - Rework the Rust tests to use `BTreeMap` instead of
   `TypedCollection`.