storage: gently remove ceased status
The design of the ceased status was predicated on definite errors
from sources being irretractable. However, in the case of
definite but not "fatal" errors, such as ingesting an value that
overflows its target type in decoding, we want to let users
retract the value to reveal the successful values.

Because some users might still have the ceased status in their status
collections, we cannot remove it wholesale. Instead, no longer issue
the ceased status and do not privilege it in the status FSM.