adapter: Fix array name resolution (#25739)
There are two ways of describing many of the builtin array types. The
named array type, for example `_int4`, and the array type with an inner
element, for example `int4[]`. During name resolution, we convert all
descriptions of the ladder to the former and save the former durably, so
`int4[]` is converted to `_int4` and we store `_int4` in the durable
catalog.

Previously, name resolution would only return the GlobalId for the type
explicitly passed in. So `int4[]` would only return the GlobalId for
`int4` and `_int4` would only return the GlobalId for `_int4`. The
GlobalIds returned from name resolution are used for various things such
as populating the `mz_object_dependencies`. As a result, if you created
an object that referenced an array type with an inner element, the set
of GlobalIds returned from name resolution could change after a restart.
For example, if you create the following table: `CREATE TABLE t (a
int4[])`, then `mz_object_dependencies` would report that the table only
depends on `int4`. In the durable catalog we store `CREATE TABLE t (a
_int4)`, which is what we parse and plan after a restart. So after a
restart, `mz_object_dependencies` would report that the same table now
only depends on `_int4`.

This commit fixes the described discrepancy by updating name resolution
so that the GlobalIds for both the named array type and the inner
element are always returned regardless of which array type variant was
passed in. So both `CREATE TABLE t (a int4[])` and `CREATE TABLE t (a
_int4)` would return the GlobalIds of both `int4` and `_int4` from name
resolution.

Fixes #21958