dbt-materialize: skip deploy_init checks if under CI and pushing to the same PR (#25744)
### Motivation

By default, `deploy_init` will fail if the schema or cluster already
exists and contains user-defined objects. However, we don't want this
check to run if someone works on a PR and pushes to the same
branch multiple times. Unforunately, GitHub actions are not expressive
enough to only run a command once. So, instead, we tag each schema and
cluster with a comment based on a CI environment variable on creation.
If set, the checks will be skipped on subsequent invocations.