compute: use a separate dyncfg for persist_source backpressure on cc
Replicas in cc sizes use lgalloc with persist, which is intended to
replace this backpressure. Use a separate dyncfg, so existing users of
the backpressure, don't rely on it when re-sizing to the cc sizes.

It's possible that this has already happened, but given that cc sizes
are prpr, we're okay with that. This new config can be used to get them
working again, but we'll have to get them to resize.