adapter: Bound ts selection in serializable
Previously, under the SERIALIZABLE isolation level, if `AS OF` was not
specified, then the query timestamp would always be advanced to the
upper. This is usually desirable and allows the user to read the most
fresh data. Static queries, such as SELECTs on a constant view, have an
empty upper and therefore have their timestamp advanced to the maximum
timestamp. For static queries with temporal filters, this results in a
predictable but sometimes confusing result. For example, consider the
following query:

   WITH dt_series AS (
       SELECT generate_series(
           date_trunc('day', date '2024-01-01'),
           date_trunc('day', date '2024-01-03'),
           '1 day'::interval
       ) AS dt
       UNION ALL
       SELECT generate_series(
           date_trunc('day', date '30000-01-01'),
           date_trunc('day', date '30000-01-03'),
           '1 day'::interval
       ) AS dt
   )
   SELECT dt FROM dt_series WHERE dt <= mz_now()

A user may think that the results under SERIALIZABLE isolation level
would not include dates in the year 30,000, given that the query was
executed before the year 30,000. However, this was not the case because
the query is static, the timestamp was advanced to the maximum
timestamp and included the results from the year 30,000.

This commit fixes this confusion by bounding the timestamp advancement
of SERIALIZABLE timestamp dependent queries to the EpochMillis
timestamp oracle read timestamp. That means that those queries will
never choose a timestamp larger than the EpochMillis timestamp oracle
read timestamp. Timestamp dependent queries are queries whose results
are affected by the chosen timestamp but do not belong to a specific
timeline. Such queries include but are not limited to static queries
with a temporal filter. Another example of timestamp dependent queries
are constant materialize views (even ones without temporal filters).