balancerd: track per-tenant metrics
This involves changing the `stream` logic a bit because its error handling was
wrong. Previously it would return an invalid password error if any error from
the stream fn occurred. However an error could occur after the client
successfully connects, and exits by terminating the connection (which psql does
if you press ctrl+D). It's wrong to return a password error in that case.

Instead, rely on the already existing run function to return that error (no
change here), and ignore the error. If the stream does init, then track the
bytes exchanged per tenant.