persist_source: wire up a start_signal
This commit adds a `start_signal` argument to the `persist_source` (and
`shard_source`) operators. A `start_signal` is a future which the
`persist_source` will await before it starts ingesting any data.

It is intended as a quick solution for suspending `persist_source`s,
which is needed for sequential dataflow hydration in compute. We plan to
remove it again once we have removed the `persist_source`-internal
backpressure and introduced a "progress control" input that can be used
for this purpose instead.