compute/persist_sink: efficient handling of future updates
This commit makes the compute `persist_sink` more efficiently handle
updates for future times by moving received updates into a stash first
and then moving them into the `correction` buffer only when the input
frontier has passed their times. The result is that we can ignore
updates for future times in the process of determining which updates
need to be sinked to persist. This has the potential for avoiding a
large amount of unnecessary work, particularly in dataflows that contain
temporal filters.