dyncfg: replace ConfigType::to_val with Into<ConfigVal>
Replace `ConfigType::to_val` with the requirement that all configuration
types implement `Into<ConfigVal>`. The signature of `to_val` was
identical to `Into<ConfigVal>`, and having the separate trait allows
methods like `add_dynamic` to express a weaker constraint. Specifically,
`add_dynamic`'s value can be any type that is convertible to a
`ConfigVal`, not just types that implement `ConfigType`. In particular,
this allows passing a `ConfigVal` directly to `add_dynamic`.