dyncfg: allow waiting for config updates (including atomically)
This commit allows dyncfg users to wait for config updates. It also
ensures waiters only seee updates *atomically*, where the unit of
atomicity is a batch of updates in `ConfigUpdates`.

The implementation is as surgical as possible. The common case or
reading a config value nonatomically is unchanged. The new
API is limited to two new methods on `ConfigSet`:

  * `ConfigSet::wait_for` runs a closure in response to config updates
    until that closure returns true.

  * `ConfigSet::with_updates_frozen` runs a closure while ensuring that
    no config batch applications are in flight, ensuring that the
    closure does not observe the partial application of updates in a
    config batch.

The new APIs will permit using dyncfg in situations where there is
interplay between multiple parameters and it is essential that updates
to those parameters are observed atomically.

The immediate need is ensuring that persist PubSub is not enabled before
we learn about the timeouts to configure for the PubSub connection
(#23869). This will be wired up in the next commit.