catalog: Flatten directory structure
This commit is purely code movement that flattens the directory
structure of the mz_catalog crate. Previously, there were multiple
implementations of the durable catalog which were all kept in their own
modules. Now that there's only a single implementation, the module
structure can be flattened which hopefully makes the crate easier to
navigate.