persist: store sum of diffs on HollowBatchPart
Begin storing the sum of diffs of all updates in a hollow part in
metadata. This sets up an optimization for `COUNT(*)`, essentially a
special-case of projection pushdown, which could use this info to
entirely avoid fetching batches with an upper less than the as_of.

The sum of diffs also happens to be a much better estimate of
cardinality for the cost based optimizer than the currently used
num_updates, but using it there is left for a followup.

Touches #20618