catalog: Simplify pre-open fencing (#25898)
Previously, the durable catalog could not be fenced while opening before
any data had been returned to a caller. This was confusing to reason
about and had an awkward confusing implementation. This behavior was
never intentionally designed and was just emergent from the stash
implementation. In order for the shadow catalog not to complain, the
persist catalog had to emulate this behavior. Now that the stash catalog
has been removed, this commit simplifies this behavior so that the
durable catalog can be fenced at any point, event before returning data
to a caller.