persist-txn: write down data shard schemas in registration records
This allows persist-txn to be self-sufficient with schemas, resolving
the cursed TODO to remove the (meaningless) default schemas passed in to
TxnsHandle on construction.

The key and value schema for each data shard is written down in the
record that registers it. Since we forget all data shards and
re-register them as part of boot, this means we (hopefully) only need
the backward compatibility shim for one release.

The production impl of TxnsCodec can sniff out the old format records by
the length of the Row (3 vs 5).