adapter,testdrive: respect overridden defaults when validating catalog
When Testdrive validates the in-memory state of the catalog against its
on disk state, in needs to take any system parameter defaults that were
overridden by the `--system-parameter-default` environmentd argument
into consideration.

The approach taken here is to have the internal /catalog/dump return
system parameter defaults as part of the catalog state. They're then
readily available to Testdrive when it loads the on-disk catalog.