compute/persist_sink: more eager batch writing
This commit makes the `write_batches` operator of the compute
`persist_sink` more eager. It does so by ceasing to gate writing of
batches by this condition:

```
!PartialOrder::less_equal(&desired_frontier, &persist_frontier)
```

(In words: Only write batches for inflight batch descriptions if the
persist frontier is less than the desired frontier.)

This condition isn't very useful. If `persist_frontier >=
desired_frontier` it means that we won't be able to successfully append
a batch, but we might still have batch descriptions for times before the
current `persist_frontier` and the only way to get rid of those in the
current implementation is writing their batches and attempting (and
failing) an append. With the above condition in place we will do so
when/if the desired frontier manages to advance past the persist
frontier, but there is no reason we shouldn't do it immediately.

We'll probably want to further refine this to ignore batch descriptions
that have no chance of being successfully appended, but that is a
larger, more risky change.